refactor: move environment v1 code to dedicated v1 package
