+++ cmd/monaco/convert/convert.go
@@ -18,13 +18,13 @@ import (
 	"fmt"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/errutils"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/log"
+	environmentv1 "github.com/dynatrace/dynatrace-configuration-as-code/pkg/environment/v1"
 	"path"
 	"path/filepath"
 
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/api"
 	configv2 "github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/converter"
-	environmentv1 "github.com/dynatrace/dynatrace-configuration-as-code/pkg/environment"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/manifest"
 	projectv1 "github.com/dynatrace/dynatrace-configuration-as-code/pkg/project/v1"
 	projectv2 "github.com/dynatrace/dynatrace-configuration-as-code/pkg/project/v2"
